add ruleset settings
