+++ osu.Game.Rulesets.Osu/OsuSettings.cs
@@ -1,0 +1,33 @@
+﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Game.Configuration;
+using osu.Game.Overlays.Settings;
+
+namespace osu.Game.Rulesets.Osu
+{
+    public class OsuSettings : SettingsSubsection
+    {
+        protected override string Header => "osu";
+
+        [BackgroundDependencyLoader]
+        private void load(OsuConfigManager config)
+        {
+            Children = new Drawable[]
+            {
+                new SettingsCheckbox
+                {
+                    LabelText = "Snaking in sliders",
+                    Bindable = config.GetBindable<bool>(OsuSetting.SnakingInSliders)
+                },
+                new SettingsCheckbox
+                {
+                    LabelText = "Snaking out sliders",
+                    Bindable = config.GetBindable<bool>(OsuSetting.SnakingOutSliders)
+                },
+            };
+        }
+    }
+}
