+++ osu.Game/Overlays/Settings/Sections/GameplaySection.cs
@@ -1,9 +1,13 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
+using osu.Framework.Allocation;
 using osu.Framework.Graphics;
+using osu.Game.Database;
 using osu.Game.Graphics;
 using osu.Game.Overlays.Settings.Sections.Gameplay;
+using osu.Game.Rulesets;
+using System.Linq;
 
 namespace osu.Game.Overlays.Settings.Sections
 {
@@ -19,6 +23,17 @@ namespace osu.Game.Overlays.Settings.Sections
                 new GeneralSettings(),
                 new SongSelectSettings(),
             };
+        }
+
+        [BackgroundDependencyLoader]
+        private void load(RulesetDatabase rulesets)
+        {
+            foreach(Ruleset ruleset in rulesets.AllRulesets.Select(info => info.CreateInstance()))
+            {
+                SettingsSubsection section = ruleset.CreateSettings();
+                if (section != null)
+                    Add(section);
+            }
         }
     }
 }
